Move search events to a commander
